refactor: fold prosody parse helpers and restore poem tree lines
- Merge footPositionFromLines and talaiLabels into linkageOverview; update barrel
- Rebuild lines from ParseResult.poem (linguistic_words/words) when present
- Merge former talaiLabels tests into linkageOverview test
