+++ src/lib/__tests__/linkageOverview.test.ts
@@ -1,6 +1,10 @@
 import { describe, expect, it } from 'vitest'
 
-import { buildLinkageOverviewRows } from '#/lib/prosody/parse/linkageOverview'
+import {
+  bondDisplayLabel,
+  buildLinkageOverviewRows,
+  linkageRowsByFromFoot,
+} from '#/lib/prosody/parse/linkageOverview'
 import type { ParsedFoot, ParsedLine, ParsedPoem } from '#/types/parsedPoem'
 
 function foot(g: number, text = 'x'): ParsedFoot {
@@ -56,5 +60,65 @@ describe('buildLinkageOverviewRows', () => {
     expect(rows[0].fromLine1).toBe(1)
     expect(rows[0].toLine1).toBe(2)
     expect(rows[0].presentationTalaiType).toBe('நேரொன்றிய ஆசிரியத்தளை')
+  })
+
+  it('indexes linkage rows by from_foot', () => {
+    const rows = buildLinkageOverviewRows({
+      original_text: '',
+      metre_type: '—',
+      letter_count: 0,
+      vikalpa_count: 0,
+      syllables: [],
+      lines: [
+        { line_class: '—', feet: [foot(0), foot(1)] },
+        { line_class: '—', feet: [foot(2)] },
+      ],
+      linkage: [
+        {
+          from_foot: 0,
+          to_foot: 1,
+          linkage_type: 'VenTalai',
+          linkage_special_type: 'IyarcirVenTalai',
+          is_valid: true,
+        },
+      ],
+    })
+    const m = linkageRowsByFromFoot(rows)
+    expect(m.get(0)).toBeDefined()
+    expect(m.get(0)?.edge.from_foot).toBe(0)
+  })
+
+  it('bondDisplayLabel prefers presentation talai string', () => {
+    const rows = buildLinkageOverviewRows({
+      original_text: '',
+      metre_type: '—',
+      letter_count: 0,
+      vikalpa_count: 0,
+      syllables: [],
+      lines: [{ line_class: '—', feet: [foot(0), foot(1)] }],
+      linkage: [
+        {
+          from_foot: 0,
+          to_foot: 1,
+          linkage_type: 'VenTalai',
+          linkage_special_type: 'VencirVenTalai',
+          is_valid: true,
+        },
+      ],
+      presentation: {
+        feet: [],
+        talai: [
+          {
+            from: 0,
+            to: 1,
+            from_line: 0,
+            to_line: 0,
+            talai_type: 'custom தளை label',
+            is_valid: true,
+          },
+        ],
+      },
+    })
+    expect(bondDisplayLabel(rows[0])).toBe('custom தளை label')
   })
 })
